% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{australia_birds}
\alias{australia_birds}
\title{The presence-absence data of birds from Australia
A dataset of bird presence absence data from Australia}
\format{A list comprising of two elements. One is 0/1 absence/presence matrix
of birds in the Australian continent - with sites along the rows and the 
bird species along the columns, with each entry being 0/1 based on
if the bird species is present in that site or not.  the other is a matrix
with equal number of rows as the presence -absence matrix with 2 columns 
representing the latitudes and longitudes.}
\usage{
australia_birds
}
\description{
The presence-absence data of birds from Australia
A dataset of bird presence absence data from Australia
}
\keyword{datasets}
