% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{australia_model}
\alias{australia_model}
\title{The ecostructure_fit() model output on Australian birds
A list of output from K=6 GoM model run on Australian bird presence absence.}
\format{A list comprising of an omega and theta matrices obtained from
\code{ecostructure_fit()} on \code{australian_birds} data for K=6}
\usage{
australia_model
}
\description{
The ecostructure_fit() model output on Australian birds
A list of output from K=6 GoM model run on Australian bird presence absence.
}
\keyword{datasets}
