% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockStructure.R
\name{BlockStructure}
\alias{BlockStructure}
\title{Ecological Block Struture plot with ggplot2 package}
\usage{
BlockStructure(omega, blocker_metadata, order_metadata,
  palette = RColorBrewer::brewer.pal(8, "Accent"), yaxis_label = "Y",
  split_line = list(split_lwd = 1, split_col = "white"),
  axis_tick = list(axis_ticks_length = 0.1, axis_ticks_lwd_y = 0.1,
  axis_ticks_lwd_x = 0.1, axis_label_size = 7, axis_label_face = "bold"),
  plot_labels = TRUE, levels_decreasing = TRUE, round_off = 1,
  layout = c(1, 2), panel_title_size = 10, panel_title_font = 4,
  main_title = "Block Structure Plot")
}
\arguments{
\item{omega}{Cluster membership probabilities of each sample. Usually a
sample by cluster matrix in the Topic model output. The cluster weights
sum to 1 for each sample.}

\item{blocker_metadata}{a factor metadata used for creating blocks of Structure plots.}

\item{order_metadata}{A quantitative metadata used for ordering sites in each block.}

\item{palette}{A vector of colors assigned to the clusters. First color in
the vector is assigned to the cluster labeled as 1, and second color in the
vector is assigned to the cluster labeled as 2, etc. The number of colors
must be the same or greater than the number of clusters. The clusters not
assigned a color are filled with white in the figure. In addition, the
recommended choice of color palette here is RColorBrewer, for instance
RColorBrewer::brewer.pal(8, "Accent") or RColorBrewwer::brewer.pal(9, "Set1").}

\item{yaxis_label}{Axis label for the samples.}

\item{split_line}{Control parameters for line splitting the batches in the
plot.}

\item{axis_tick}{Control parameters for x-axis and y-axis tick sizes.}

\item{plot_labels}{A logical parameter, if TRUE the function plots the axis labels.}

\item{levels_decreasing}{if TRUE, the sites in the block are ordered in decreasing order
of magnitude of ordering metadata. Default is TRUE.}

\item{round_off}{The ordering metadata is rounded off to these many digits.}

\item{layout}{The graph layout for plotting the block Structure output.}

\item{panel_title_size}{The size of the title for layout panel}

\item{panel_title_font}{The font of the title for the layout panel}

\item{main_title}{the title of the Block Structure model plot.}
}
\value{
Plots the Block Structure plot visualization of the GoM model
}
\description{
Make a Structure Plot split into ecological blocks and ordered
in each block by an ordering metadata
}
\examples{

library(HimalayanBirdsAbundance)
data("HimalayanBirdsAbundance")
library(Biobase)
new_counts <- t(exprs(HimalayanBirdsAbundance));
metadata <- pData(HimalayanBirdsAbundance);
elevation_metadata=metadata$Elevation;
east_west_dir = metadata$WorE;
topic_clus <- maptpx::topics(new_counts, K=2, tol=0.1)
omega <- topic_clus$omega
BlockStructure(omega, blocker_metadata = east_west_dir,
              order_metadata = elevation_metadata,
              yaxis_label = "Elevation",
              levels_decreasing = FALSE)

}

