% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dispersion_field_ex}
\alias{dispersion_field_ex}
\title{A dispersion field matrix example - used for creating a map}
\format{a dispersion field example created by \code{dispersion_fields_create}
in matrix format.}
\usage{
dispersion_field_ex
}
\description{
A dispersion field matrix example - used for creating a map
}
\keyword{datasets}
