% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_fields_to_maps.R
\name{dispersion_fields_to_maps}
\alias{dispersion_fields_to_maps}
\title{Create geographical heatmaps from dispersion field list}
\usage{
dispersion_fields_to_maps(dispersion.field, global_shapefile,
  raster_latlim = c(5, 50), raster_longlim = c(50, 120),
  proj = CRS(" +proj=longlat +ellps=WGS84"), color_ramp = c("darkseagreen3",
  "orange", "red"))
}
\arguments{
\item{dispersion.field}{The list of dispersion field matrices per site 
generated by \code{dispersion_fields_create} function.}

\item{global_shapefile}{Global shapefile to read shape lines for country 
boundaries in background.}

\item{raster_latlim}{The latitudinal range of the dispersion fields}

\item{raster_longlim}{the longitudinal range of the dispersion fields}

\item{proj}{the CRS projection of the global shapfile to apply to each 
dispersion field matrix}

\item{color_ramp}{Color palette to be for the heat map}
}
\description{
Creates geographical heatmaps depicting global dispersion field
patterns of the different sites in the local data frame from the dispersion
field list generated by the \code{dispersion_fields_create} and a global
shapefile to sketch the background.
}
\examples{
data("dispersion_field_ex")
proj <- CRS(' +proj=longlat +ellps=WGS84')
global_shapefile <- sf::st_read(system.file("extdata",
                    "ne_110m_coastline",
                    "ne_110m_coastline.shp",
                     package = "ecostructure"), 
                    quiet = TRUE)
maps <- dispersion_fields_to_maps(dispersion_field_ex, global_shapefile )

}
