% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersionFieldToCounts.R
\name{DispersionFieldToCounts}
\alias{DispersionFieldToCounts}
\title{Creates global site by latitudinal-longitudinal grid cell
dispersion field abundance counts.}
\usage{
DispersionFieldToCounts(dispersion.field)
}
\arguments{
\item{dispersion.field}{Dispersion field list generated by
\code{CreateGlobalAbundanceMaps} function.}
}
\value{
The global site by grid cell abundance counts matrix.
}
\description{
Create global abundance counts matrix by merging the dispersion
fields from the list of dispersion field maps generated by
\code{CreateGlobalAbundanceFields} function.
}
\examples{

dispersion.field <- get(load("dispersion_field_mat.rda"))
counts <- DispersionFieldToCounts(dispersion.field)


}
\keyword{abundance}
\keyword{counts}
\keyword{dispersion}
\keyword{field,}
\keyword{global}

