% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_motif.R
\name{dsp_motif}
\alias{dsp_motif}
\title{Creates heatmap for a geographic motif}
\usage{
dsp_motif(theta_mat, dsp_fld_res = 8, color_ramp = c("black",
  "darkseagreen3", "orange", "red"), raster_latlim = c(5, 50),
  raster_longlim = c(50, 120), outline_col = "white",
  proj = "+proj=longlat +ellps=WGS84")
}
\arguments{
\item{theta_mat}{The theta output from the model fit from 
\code{ecos_fit}}

\item{dsp_fld_res}{The resolution of the dispersion field.}

\item{color_ramp}{Color palette to be used for the heat map.}

\item{raster_latlim}{If disp_fld_list is a list of matrices, the latitudinal range 
of the dispersion fields must be provided as a vector.}

\item{raster_longlim}{If disp_fld_list is a list of matrices, the longitudinal range 
of the dispersion fields must be provided as a vector.}

\item{outline_col}{The outline color of the projected map.}

\item{proj}{A character string of projection arguments; the arguments must be entered 
exactly as in the PROJ.4 documentation. To be evaluated by sp::CRS(). 
default '+proj=longlat +ellps=WGS84'.}
}
\value{
Returns a heatmap plot, similar to the \code{dsp_plot_map} function
        for the theta matrix using the specified raster,
}
\description{
Takes the motif fitted from \code{ecos_fit} and plots the theta
component as a heatmap across space using suitable raster information.
}
\examples{

data("example_theta")
himalayan_geo_motifs <- dsp_motif(example_theta, 
                                 color_ramp = c("black", "darkseagreen3",
                                                "orange","red"),
                                dsp_fld_res = 8)
himalayan_geo_motifs$motif_maps[[1]]

}
\keyword{dispersion}
\keyword{field,}
\keyword{maps}
\keyword{motif,}
