% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_plot_map.R
\name{dsp_plot_map}
\alias{dsp_plot_map}
\title{Create geographical heatmaps from a dispersion field list}
\usage{
dsp_plot_map(disp_fld_list, scale = "count", color_ramp = c("darkseagreen3",
  "orange", "red"), raster_latlim = c(5, 50), raster_longlim = c(50, 120),
  outline_col = "black", proj = "+proj=longlat +ellps=WGS84")
}
\arguments{
\item{disp_fld_list}{The list of dispersion fields generated by the
\code{dsp_create_from_survey} function. Objects in the list
can be RasterLayers, it which case proj and boundings 
from the rasters are used for plotting, or matrices, 
in which case proj and boundings must be provided.}

\item{scale}{Type of scale to be used for the legend. One of 
c("count","percentage"). Defaults "count".}

\item{color_ramp}{Color palette to be used for the heat map.}

\item{raster_latlim}{If disp_fld_list is a list of matrices, the latitudinal range 
of the dispersion fields must be provided as a vector.}

\item{raster_longlim}{If disp_fld_list is a list of matrices, the longitudinal range 
of the dispersion fields must be provided as a vector.}

\item{outline_col}{The outline color of the projected map.}

\item{proj}{A character string of projection arguments; the arguments must be entered 
exactly as in the PROJ.4 documentation. To be evaluated by sp::CRS(). 
default '+proj=longlat +ellps=WGS84'.}
}
\value{
Returns a heatmap of the dispersion field using the specified raster
}
\description{
Creates simple geographical heatmaps for easily visualizing 
 dispersion fields. Uses a dispersion field list generated by 
 the \code{dsp_create_from_survey} and global shapefiles 
 of coastlines and country boundaries to sketch the background.
}
\examples{
data("dispersion_field_ex")
maps <- dsp_plot_map(dispersion_field_ex, 
                     scale = "percentage",
                     raster_latlim = c(5,50),
                     raster_longlim = c(50,120))
maps[[1]]

}
