% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_to_matrix.R
\name{dsp_to_matrix}
\alias{dsp_to_matrix}
\title{Creates global site by latitudinal-longitudinal grid cell
dispersion field abundance counts.}
\usage{
dsp_to_matrix(dispersion.field)
}
\arguments{
\item{dispersion.field}{Dispersion field list generated by
\code{CreateGlobalAbundanceMaps} function.}
}
\value{
The global site by grid cell abundance counts matrix.
}
\description{
Create global abundance counts matrix by merging the dispersion
fields from the list of dispersion field maps generated by
\code{CreateGlobalAbundanceFields} function.
}
\examples{

data("dispersion_field_ex")
mat <- dsp_to_matrix(dispersion_field_ex)


}
\keyword{abundance}
\keyword{counts}
\keyword{dispersion}
\keyword{field,}
\keyword{global}
