% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecos_fit.R
\name{ecos_fit}
\alias{ecos_fit}
\title{Fitting Grade of Membership model for clustering into communities}
\usage{
ecos_fit(
  dat,
  max_dat = NULL,
  K,
  tol = 0.1,
  num_trials = 1,
  fit_control = list()
)
}
\arguments{
\item{dat}{input data matrix of samples along rows and features along sites,
with each entry a 0/1 signififying presence/absence or counts of
abundances.}

\item{K}{The number of clusters to fit}

\item{tol}{The tolerance level of the model.}

\item{num_trials}{Number of EM runs from different starting points. This is'
key for picking the best fit model across multiple runs.}

\item{fit_control}{The control parameters for the model.}
}
\value{
Returns a model fit with \code{omega} as cluster membership 
        probabilities and \code{theta} as cluster feature matrix.
}
\description{
A grade of membership model for clustering a site by features
data - which could be presence-absence or counts abundance data of species
in the sites.
}
\examples{

data("himalayan_birds")
species_abundance_counts <- t(exprs(himalayan_birds));
fit <- ecostructure_fit(species_abundance_counts, K = 2, tol = 0.1)
species_pa_counts <- species_abundance_counts
species_pa_counts[species_pa_counts >=1] = 1
fi2 <- ecos_fit(species_pa_counts, K = 2, tol = 0.1)

}
