% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecos_fit_dev.R
\name{ecos_fit_dev}
\alias{ecos_fit_dev}
\title{Fitting Grade of Membership model for clustering into communities}
\usage{
ecos_fit_dev(
  dat,
  max_dat = NULL,
  K,
  sparse = F,
  num_iter_main = 100,
  num_iter_refine = 100,
  fit_control_init = list(),
  fit_control_main = list(numiter = 4),
  fit_control_refine = list(numiter = 4, extrapolate = TRUE)
)
}
\arguments{
\item{dat}{input data matrix of samples along rows and features along sites,
with each entry a 0/1 signifying presence/absence or counts of
abundances.}

\item{K}{The number of clusters to fit}

\item{tol}{The tolerance level of the model.}

\item{num_trials}{Number of EM runs from different starting points. This is'
key for picking the best fit model across multiple runs.}

\item{fit_control}{The control parameters for the model.}
}
\value{
Returns a model fit with \code{omega} as cluster membership
probabilities and \code{theta} as cluster feature matrix.
}
\description{
A grade of membership model for clustering a site by features
data - which could be presence-absence or counts abundance data of species
in the sites.
}
\examples{

data("himalayan_birds")
species_abundance_counts <- t(exprs(himalayan_birds))
fit <- ecostructure_fit(species_abundance_counts, K = 2, tol = 0.1)
species_pa_counts <- species_abundance_counts
species_pa_counts[species_pa_counts >= 1] <- 1
fi2 <- ecos_fit(species_pa_counts, K = 2, tol = 0.1)

}
