% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecos_nullmodel.R
\name{ecos_nullmodel}
\alias{ecos_nullmodel}
\title{Null models for ecostructure_fit clustering validation}
\usage{
ecos_nullmodel(
  counts,
  K,
  tol = 0.1,
  null.model = c("frequency", "richness", "independentswap", "trialswap"),
  iter_fill = 100,
  iter_randomized = 30,
  option,
  plot = TRUE
)
}
\arguments{
\item{counts}{The counts matrix (N x G): N- the number of samples (sites), 
G- number of features (bird species)}

\item{K}{The number of clusters to fit}

\item{tol}{The tolerance limit of the \code{ecos_fit}.}

\item{null.model}{The type of nullmodel used (similar to the
picante::randomizeMatrix() function argument in 
picante package)}

\item{iter_fill}{The number of swaps/fills in each randomized matrix build}

\item{iter_randomized}{The number of randomization matrices generated}

\item{plot}{If TRUE, plots density function of log Bayes factor over
the randomized iterations}
}
\value{
Returns a list with
       \item{BF.obs}{log Bayes Factor for the observed counts with 
       K=2 against the null with no clusters}
       \item{BF.rand}{a vector of log BF for each randomized count 
       matrix with K=2 against the null with no clusters}
       \item{pval}{the p-value of the observed log Bayes factor against the
       ones from randomized matrices}
}
\description{
Generate randomized matrix of counts 
given the observed data matrix for a particular null model, 
run \code{ecos_fit} on these null matrices and compare the fit
on null model data with that on the observed data. Used for validating 
the clustering.
}
\examples{

data("himalayan_birds")
species_abundance_counts <- t(exprs(himalayan_birds));
out <- ecos_nullmodel(species_abundance_counts, K=2, 
                iter_randomized=5, option = "BF")
out2 <- ecos_nullmodel(species_abundance_counts, K=2, 
                iter_randomized=5, ption = "BIC")              
  
}
