% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecostructure_blocks.R
\name{ecostructure_blocks}
\alias{ecostructure_blocks}
\title{Ecological Block Struture plot with ggplot2 package}
\usage{
ecostructure_blocks(omega, blocker_metadata, order_metadata,
  palette = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2",
  "#D55E00", "#CC79A7"), structure_control = list(), layout)
}
\arguments{
\item{omega}{Matrix of Cluster membership probabilities of each sample
obtained from \code{ecostructure_fit()}. The row sums 
sum to 1 for each sample.}

\item{blocker_metadata}{a factor metadata used for creating blocks of 
Structure plots.}

\item{order_metadata}{A quantitative metadata used for ordering sites 
in each block.}

\item{palette}{A vector of colors assigned to the clusters. First color in
the vector is assigned to the cluster labeled as 1, and second
color in the vector is assigned to the cluster labeled as 2, 
etc. The number of colors must be the same or greater than the
number of clusters. The recommended choice of color palette 
here is RColorBrewer, for instance 
RColorBrewer::brewer.pal(8, "Accent") or 
RColorBrewwer::brewer.pal(9, "Set1").}

\item{structure_control}{Control parameters for the Block Structure plot.
Fixes the title, axis labels, tick sizes, splitting
line characteristics and panel orientation.}

\item{layout}{The graph layout for plotting the Block Structure output.
Is missing, automatically determined as a square 
configuration.}
}
\value{
Plots the Block Structure plot visualization of the GoM model
}
\description{
Make a Structure Plot split into ecological blocks by a blocker
metadata and ordered in each block by an ordering metadata
}
\examples{

library(HimalayanBirdsAbundance)
data("HimalayanBirdsAbundance")
library(Biobase)
new_counts <- t(exprs(HimalayanBirdsAbundance));
metadata <- pData(HimalayanBirdsAbundance);
elevation_metadata=metadata$Elevation;
east_west_dir = metadata$WorE;
topic_clus <- maptpx::topics(new_counts, K=2, tol=0.1)
omega <- topic_clus$omega
BlockStructure(omega, blocker_metadata = east_west_dir,
              order_metadata = elevation_metadata,
              yaxis_label = "Elevation",
              levels_decreasing = FALSE)

}
