% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecostructure_fit.R
\name{ecostructure_fit}
\alias{ecostructure_fit}
\title{Fitting Grade of Membership model for clustering into communities}
\usage{
ecostructure_fit(dat, max_dat = NULL, K, tol = 0.1, fit_control = list())
}
\arguments{
\item{dat}{input data matrix of samples along rows and features along sites,
with each entry a 0/1 signififying presence/absence or counts of
abundances.}

\item{K}{The number of clusters to fit}

\item{tol}{The tolerance level of the model.}

\item{fit_control}{The control parameters for the model.}
}
\value{
Returns a model fit with \code{omega} as cluster membership 
        probabilities and \code{theta} as cluster feature matrix.
}
\description{
A grade of membership model for clustering a site by features
data - which could be presence-absence or counts abundance data of species
in the sites.
}
