% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecostructure_plot_pie.R
\name{ecostructure_plot_pie}
\alias{ecostructure_plot_pie}
\title{Produces ecological pie chart representations of ecostructure_fit model}
\usage{
ecostructure_plot_pie(omega = NULL, coords = NULL, bgmap_path = NULL,
  adjust = FALSE, thresh = 0.7, long_lim = c(-180, 180),
  lat_lim = c(-60, 90), coastline_lwd = 10, intensity = 1, radius = 0.5,
  color = c("dodgerblue2", "#E31A1C", "green4", "#6A3D9A", "#FF7F00", "black",
  "gold1", "skyblue2", "#FB9A99", "palegreen2", "#CAB2D6", "#FDBF6F", "gray70",
  "khaki2", "maroon", "orchid1", "deeppink1", "blue1", "steelblue4",
  "darkturquoise", "green1", "yellow4", "yellow3", "darkorange4", "brown",
  "red", "cornflowerblue", "cyan", "brown4", "burlywood", "darkgoldenrod1",
  "azure4", "green", "deepskyblue", "yellow", "azure1"), pie_control = list(),
  image_width = 1000, image_height = 800, path = "geostructure_plot.tiff")
}
\arguments{
\item{omega}{Matrix of Cluster membership probabilities of each sample
obtained from \code{ecostructure_fit()}. The row sums 
sum to 1 for each sample.}

\item{coords}{a matrix of latitude and longitude entries for each site (row)
of the \code{omega} matrix.This matrix has as many rows as 
the \code{omega} matrix and two columns for the latitude and 
longitude.}

\item{bgmap_path}{The path to the shapefile to be used for plotting the 
map data. One can choose a shapefile of their own 
preference of resolution. If missing, we use a default 
background map - which is a coarse one.}

\item{adjust}{A Boolean, indicating whether to adjust for the richness cluster.
Defaults to TRUE.}

\item{thresh}{a numeric value between 0 and 1 indicating the level at which
to threshold the richness cluster. Sites represented by member-
ship probability above this threshold of the richness cluster
are removed.}

\item{long_lim}{The limits for the longitude in the map plot.
The default is \code{c(-180,180)} and removes Antarctica 
from the map.}

\item{lat_lim}{The limits for the latitude in the map plot.
The default is \code{c(-60,90)} and removes Antarctica 
from the map.}

\item{coastline_lwd}{The line width for the coastline plotted in the map.
Defaults to 10.}

\item{intensity}{The intensity of the colors for the pie chart representation.
Lies between 0 and 1. Defaults to 1.}

\item{radius}{The radius of the pie charts. Defaults to 0.5}

\item{color}{the vector of colors - of same size or more than the number 
of clusters. If missing, we use a default set of colors- 
chosen so that they are distinct from one another.}

\item{pie_control}{The list of control parameters to be passed into the 
\code{add.pie} function of the package maptools.}

\item{image_width}{The width of the image output. Defaults to 1000.}

\item{image_height}{The height of the image output. Defaults to 800.}

\item{path}{The path where the output image is saved.}
}
\value{
Returns a pie chart representation of the memberships obtained 
        from \code{ecostructure_fit} on a global species 
        presence-absence/counts data
}
\description{
Make a pie chart representation of the membership probabilities
\code{omega} output from the \code{ecostructure_fit} model.
}
\examples{
 data("australia_birds")
 data("australia_model")
 ecostructure_plot_pie(omega = australia_model$omega,
                     coords = australia_birds$latlong, 
                     long_lim = c(110,160),
                     lat_lim = c(-50,-10),
                     color= c("orange", "red", "yellow", "deepskyblue", 
                              "chartreuse", "blue"))



}
