% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecostructure_prepare_by_phylo.R
\name{ecostructure_prepare_by_phylo}
\alias{ecostructure_prepare_by_phylo}
\title{Prepare counts data for phylogenetically collapsed species}
\usage{
ecostructure_prepare_by_phylo(dat, tree, collapse_at)
}
\arguments{
\item{tree}{The phylogenetic tree structure on the species or the 
column features of the counts matrix. Should be of the form
\code{.tre}.}

\item{collapse_at}{The point on the tree at which to collapse the tree}

\item{counts}{The abundance counts matrix with samples/sites along the rows 
and the species/features along the columns.}
}
\value{
Returns a counts matrix with the features being the 
phylogenetically collapsed species from the original data based on the 
tree structure.
}
\description{
Starting from a taxonomic species abundance matrix, produces
a new matrix, with features now collpased vesions of the species where the
collapsing is done by a phylogenetic tree structure.
}
