% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecostructure_prepare_by_trait.R
\name{ecostructure_prepare_by_trait}
\alias{ecostructure_prepare_by_trait}
\title{Prepare counts data for trait-collapsed species}
\usage{
ecostructure_prepare_by_trait(counts, traits, prop_div = 0.3)
}
\arguments{
\item{counts}{The abundance counts matrix with samples/sites along the rows 
and the species/features along the columns.}

\item{traits}{A matrix with species along the rows and traits along
the columns.}

\item{prop_div}{The proportion of the original counts matrix diversity at which
to cut the dendrogram.}
}
\value{
The output is a counts matrix with the columns corresponding to the 
species bins - each bin designated here by the first species entry into the bin.
}
\description{
Starting from a taxonomic species abundance matrix, produces
a new matrix, with features now collpased vesions of the species where the
collapsing is done by trait level information. Binning is done on the species
based on their trait level similarity- where similarity is determined by
a dendogram produced by hierarchical clustering.
}
\examples{

data <- get(load(system.file("extdata", "HimalayanBirdsData.rda", package = "ecostructure")))
species_metadata <- pData(featureData(data))
taxonomic_counts <- t(exprs(data))
bill_traits <- as.matrix(dist(scale(species_metadata[,c(1:3)])))
bill_trait_clust <- trait_cluster(counts = taxonomic_counts, traits = bill_traits, prop_div=0.3)

}
