% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{him_birds_cn}
\alias{him_birds_cn}
\title{Abundances of Himalayan bird species in local communities across elevations
A dataset of bird species abundances in local communities in the Himalayas}
\format{
A matrix object with 38 sites (rows) and
304 species (columns)
}
\usage{
him_birds_cn
}
\description{
Abundances of Himalayan bird species in local communities across elevations
A dataset of bird species abundances in local communities in the Himalayas
}
\keyword{datasets}
