% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{him_birds_md}
\alias{him_birds_md}
\title{Morphological and functional metadata of Himalayan birds species
A dataset of morphological and functional traits of birds censused acrross}
\format{
An ExpressionSet object with 38 sites (rows) and
304 species (columns) including phenoData which is site metadata and
featureData which consists of feature metadata
}
\usage{
him_birds_md
}
\description{
Morphological and functional metadata of Himalayan birds species
A dataset of morphological and functional traits of birds censused acrross
}
\keyword{datasets}
