% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{himalayan_birds}
\alias{himalayan_birds}
\title{Local abundance data of Himalayan bird species with metadata
A dataset of bird abundance counts with site and feature metadata}
\format{
An ExpressionSet object with 38 sites (rows) and 
304 species (columns) including phenoData which is site metadata and 
featureData which consists of feature metadata
}
\usage{
himalayan_birds
}
\description{
Local abundance data of Himalayan bird species with metadata
A dataset of bird abundance counts with site and feature metadata
}
\keyword{datasets}
