% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_climate_raster_by_GoM.R
\name{mask_climate_raster_by_GoM}
\alias{mask_climate_raster_by_GoM}
\title{Mask a climate raster by the motif map raster}
\usage{
mask_climate_raster_by_GoM(motif_map_raster, climate_raster, quant = 0.1,
  index = 10)
}
\arguments{
\item{motif_map_raster}{The motif map raster obtained from fitting GoM model on the
species abundance counts data.}

\item{climate_raster}{the climate raster which is to be masked.}

\item{quant}{A scaling factor determining how fine the raster grid is}

\item{index}{The number of levels to be used}
}
\value{
Returns a masked climate raster.
}
\description{
Using the motif map raster obtained from applying Grade of
Membership model on the counts data, this function can mask climate raster -
including temperature, precipitation etc.
}
