% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_counts.R
\name{order_counts}
\alias{order_counts}
\title{Preparing ordered counts data for ordered topic model.}
\usage{
order_counts(local_data, ordering_vec, threshmax = 5)
}
\arguments{
\item{local_data}{The abundance counts data for site/species.}

\item{ordering_vec}{The vector used for ordering the columns of the data
matrix (species in this scenario).}

\item{cut_off_percentile}{The percentile cut off used for padding the zeros.}
}
\value{
Returns an ordered counts matrix with the species along the columns
ordered by the \item{ordering_vec}.
}
\description{
Ordering counts data based on an ordering vector and padding
zeros at appropriate locations, so that the total number of columns is a
power of 2, a constraint one needs to apply the ordered topic model.
}
