% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phylo_tree}
\alias{phylo_tree}
\title{An object of type phylo - which is a phylogenetic tree}
\format{
An object of type phylo of the 304 bird species of the Himalayas
}
\usage{
phylo_tree
}
\description{
An object of type phylo - which is a phylogenetic tree
}
\keyword{datasets}
