% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_cluster.R
\name{trait_cluster}
\alias{trait_cluster}
\title{Collapsing the species in a ecological counts data using trait based grouping}
\usage{
trait_cluster(counts, traits, prop_div = 0.3)
}
\arguments{
\item{counts}{The abundance counts data matrix with samples along the rows and the
species along the columns. The sample names are provided as row names
and the species names represent the column names.}

\item{traits}{A matrix with species along the rows. Teh species names along the rows
must match with the column names of the counts matrix input.}

\item{prop_div}{The proportion of the original counts matrix diversity at which
to cut the dendrogram.}
}
\value{
The output is a counts matrix with the columns corresponding to the clusters and
the column name corresponding to the first of the species forming the cluster. This
matrix is taken as input to the CountClust:FitGoM function.
}
\description{
Performs grouping of species in the abundance counts daya based on trait based
hierarchical clustering of the species
}
\examples{

data <- get(load(system.file("extdata", "HimalayanBirdsData.rda", package = "ecostructure")))
species_metadata <- pData(featureData(data))
taxonomic_counts <- t(exprs(data))
bill_traits <- as.matrix(dist(scale(species_metadata[,c(1:3)])))
bill_trait_clust <- trait_cluster(counts = taxonomic_counts, traits = bill_traits, prop_div=0.3)

}
