% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_create_from_gdb.R
\name{dsp_create_from_gdb}
\alias{dsp_create_from_gdb}
\title{Create dispersion fields and a presence absence matrix from geodatabase 
       file or a shape file containing all species.}
\usage{
dsp_create_from_gdb(
  gdb_object,
  raster_resolution = 5,
  thresh = 4,
  raster_latlim = c(-90, 90),
  raster_longlim = c(-180, 180),
  species_feature = "SCINAME",
  precision = 0.25,
  species_names = NULL,
  drop = FALSE
)
}
\arguments{
\item{gdb_object}{The geodatabase or a single shapefile containing all species to be read 
in ahead of time and provided as an sf object.}

\item{raster_resolution}{The resolution of the raster in degrees.}

\item{thresh}{The number of co-occuring species required to generate a dispersion field for that map cell.}

\item{raster_latlim}{The latitudinal range of the desired dispersion fields and presence-absence matrix.}

\item{raster_longlim}{The longitudinal range.}

\item{species_feature}{The name of the feature (column) that contains the species names 
in the GIS data source.}

\item{precision}{The precision of the initial estimation provided to fasterize. 
fasterize tends to undercount small species ranges, so is required 
to keep at a low value in order detect species presences.
Most users will not need to change, but increaseing this value will 
speed up the function but with potential for error.}

\item{species_names}{If you would like to only include a subset of the species is the GIS data source
a vector of species names can be provided to filter the GIS data before processing
into a presence absence matrix and dispersion fields.}

\item{drop}{Boolean. If T sites with 0 overlapping species will be dropped from the final presence-absence matrix.
Defaults to false.}
}
\value{
Returns 2 lists and 1 matrix. In each list, each element is a
dispersion field corresponding to the species present in the 
specific site to which it corresponds. The first list "matrix" 
contains each dispersion field as a matrix. The second list "raster" returns 
each dispersion field as a raster. The matrix is a presence absence matrix with 
species as columns and each individual map cell the rows. The names of the elements 
of the lists, and the rownames for the presence-absence matrix, are assigned as 
the lat_long combos for those locations.
}
\description{
Reads a single GIS data file and generates a presence-matrix
and associated dispersion fields in a user-specified region.
}
\keyword{counts}
\keyword{data,}
\keyword{dispersion}
\keyword{field}
\keyword{local}
\keyword{site-species}
