% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_motif.R
\name{dsp_motif}
\alias{dsp_motif}
\title{Creates global site by latitudinal-longitudinal grid cell
dispersion field abundance counts.}
\usage{
dsp_motif(theta_mat, dsp_fld_res = 8, color_ramp = c("black",
  "darkseagreen3", "orange", "red"), raster_latlim = c(5, 50),
  raster_longlim = c(50, 120), outline_col = "white",
  proj = "+proj=longlat +ellps=WGS84")
}
\arguments{
\item{dispersion.field}{Dispersion field list generated by
\code{CreateGlobalAbundanceMaps} function.}
}
\value{
The global site by grid cell abundance counts matrix.
}
\description{
Create global abundance counts matrix by merging the dispersion
fields from the list of dispersion field maps generated by
\code{CreateGlobalAbundanceFields} function.
}
\examples{

}
\keyword{abundance}
\keyword{counts}
\keyword{dispersion}
\keyword{field,}
\keyword{global}
