% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige_counts.R
\name{krige_counts}
\alias{krige_counts}
\title{Kriging smoothed counts data preparation}
\usage{
krige_counts(counts, order, krige.control = list(), jitter = 1e-06)
}
\arguments{
\item{counts}{The abundance counts data matrix with samples along the rows and the
species along the columns.}

\item{order}{The ordered feature information used for computing distance information
in kriging.}

\item{krige.control}{The control parameters for kriging.}

\item{jitter}{A jitter value to be added in case of any singularity problem due to lack of
uniqueness of ordering metadata. Default is 10^-6. The user may tune it in case
of a singularity problem found.}
}
\value{
Returns a counts data with the zero count cells being interpolated by
a simple kriging predictor (using \code{SpatialExtremes} package).
}
\description{
Performs interpolation of 0 counts (treated as missing data) using
kriging over an ordered feature information \code{order}.
}
